docs/requirements.txt
sphinx-rtd-theme==2.0.0
metomi-rose
cylc-flow
cylc-sphinx-extensions.git[all]
